+++ src/MasterProduct/resources/database/migrations/2022_06_13_090835_create_master_product_variants_table.php
@@ -13,6 +13,7 @@ return new class () extends Migration {
             $table->bigIncrements('id');
             $table->unsignedBigInteger('master_product_id');
 
+            $table->string('sku');
             $table->string('name')->nullable();
             $table->text('excerpt')->nullable();