+++ Changelog.md
@@ -5,6 +5,7 @@
 ## Unreleased
 ##### 2022-06-XX
 
+- Added the Master Product module
 - Added the Adjustments module
 - Added the extended Adjustable Cart model to `Foundation` (original Cart model still there, intact)
 - Added Shipping methods